ShipMMG

Documentation for ShipMMG.

ShipMMG.Mmg3DofBasicParams

ShipMMG.Mmg3DofManeuveringParams

ShipMMG.draw_gif_result

ShipMMG.kt_model!

ShipMMG.kt_simulate

ShipMMG.kt_zigzag_test

ShipMMG.mmg_3dof_model!

ShipMMG.mmg_3dof_simulate

ShipMMG.mmg_3dof_zigzag_test

ShipMMG.simulate

ShipMMG.Mmg3DofBasicParams — Type

Basic parameters of target ship for MMG 3DOF simulation.

Arguments

L_pp::Float64: L_pp

B::Float64:

d::Float64:

x_G::Float64:

D_p::Float64:

m::Float64:

I_zG::Float64:

A_R::Float64:

η::Float64:

m_x::Float64:

m_y::Float64:

J_z::Float64:

f_α::Float64:

ϵ::Float64:

t_R::Float64:

a_H::Float64:

x_H::Float64:

γ_R_minus::Float64:

γ_R_plus::Float64:

l_R::Float64:

κ::Float64:

t_P::Float64:

w_P0::Float64:

x_P::Float64:

ShipMMG.Mmg3DofManeuveringParams — Type

Maneuvering parameters of target ship for MMG 3DOF simulation.

Arguments

k_0::Float64

k_1::Float64

k_2::Float64

R_0_dash::Float64

X_vv_dash::Float64

X_vr_dash::Float64

X_rr_dash::Float64

X_vvvv_dash::Float64

Y_v_dash::Float64

Y_r_dash::Float64

Y_vvv_dash::Float64

Y_vvr_dash::Float64

Y_vrr_dash::Float64

Y_rrr_dash::Float64

N_v_dash::Float64

N_r_dash::Float64

N_vvv_dash::Float64

N_vvr_dash::Float64

N_vrr_dash::Float64

N_rrr_dash::Float64

ShipMMG.draw_gif_result — Function

draw_gif_result(time, x, y, ψ, shape, file_path, [, fps]) -> gif

Draw the gif animation from simulation result.

ShipMMG.kt_model! — Method

kt_model!(dX, X, p, t)

KT model on DifferentialEquations.ODEProblem. Update dX.

Arguments

dX: [du, dr, dδ]

X: the initial state values. [u, r, δ].

p: the parameters and δ spline info. [K, T, spl_δ].

t: the time.

Examples

julia> K_log = 0.155  # [1/s]
julia> T_log = 80.5  # [s]
julia> u0 = 20 * (1852.0 / 3600)  # [m/s] (knot * 1852/3600)
julia> duration = 50  # [s]
julia> sampling = 1001
julia> time_list = range(0.0,stop=duration,length=sampling)
julia> Ts = 50.0
julia> δ_list = 10.0 * pi / 180.0  * sin.(2.0 * pi / Ts * time_list) # [rad]
julia> spl_δ = Spline1D(time_list, δ_list)
julia> X0 = [u0; 0.0; δ_list[1]]
julia> p = [K_log, T_log, spl_δ]
julia> prob = ODEProblem(kt_model!, X0, (time_list[1], time_list[end]), p)
julia> sol = solve(prob, Tsit5(), saveat=time_list[2] - time_list[1])

ShipMMG.kt_simulate — Method

kt_simulate(time_list, δ_list, K, T, [, u0, r0, algorithm, reltol, abstol])

Returns the KT simulation results.

Arguments

time_list: the list of simulatino time.

δ_list: the list of rudder angle [rad].

K::Float64: the K Parameter.

T::Float64: the T Parameter.

u0::Float64=0.0: the constant surge velocity.

r0::Float64=0.0: the inital rate of turn [rad/s].

algorithm=Tsit5(): the parameter of DifferentialEquations.ODEProblem.solve()

reltol=1e-8: the parameter of DifferentialEquations.ODEProblem.solve()

abstol=1e-8: the parameter of DifferentialEquations.ODEProblem.solve()

Examples

KT simulation.

julia> K_log = 0.155  # [1/s]
julia> T_log = 80.5  # [s]
julia> u0 = 20 * (1852.0 / 3600)  # [m/s] (knot * 1852/3600)
julia> duration = 50  # [s]
julia> sampling = 1001
julia> time_list = range(0.0,stop=duration,length=sampling)
julia> Ts = 50.0
julia> δ_list = 10.0 * pi / 180.0  * sin.(2.0 * pi / Ts * time_list) # [rad]
julia> time, u, r, δ = kt_simulate(time_list, δ_list, K_log, T_log, u0=u0)

ShipMMG.kt_zigzag_test — Method

kt_zigzag_test(K, T, target_δ_rad, target_ψ_rad_deviation, time_second_interval, end_time_second, [, u0, r0, ψ0, δ0, δ_rad_rate, start_time_second, algorithm, reltol, abstol])

Returns the KT simulation results.

Arguments

K::Float64: the K Parameter.

T::Float64: the T Parameter.

target_δ_rad::Float64: target rudder angle of zigzag test.

target_ψ_rad_deviation::Float64: target azimuth deviation of zigzag test.

time_second_interval::Float64: time interval of output result.

end_time_second::Float64: the end time of simulation.

u0::Float64=0.0: the initial & constant surge velocity.

r0::Float64=0.0: the initial rate of turn [rad/s].

ψ0::Float64=0.0: the initial azimuth.

δ0::Float64=0.0: the initial rudder angle.

δ_rad_rate::Float64=10.0*π/180: the change rate of rudder angle [rad/s].

start_time_second::Float64=0.0: the start time of simulation.

algorithm=Tsit5(): the parameter of DifferentialEquations.ODEProblem.solve()

reltol=1e-8: the parameter of DifferentialEquations.ODEProblem.solve()

abstol=1e-8: the parameter of DifferentialEquations.ODEProblem.solve()

Examples

KT simulation.

julia> K = 0.155
julia> T = 80.5
julia> target_δ_rad = 30.0 * π / 180.0
julia> target_ψ_rad_deviation = 10.0 * π / 180.0
julia> time_second_interval = 0.01
julia> end_time_second = 500.0
julia> u0 = 20 * (1852.0 / 3600) # [knot]
julia> δ_list, r_list, ψ_list = kt_zigzag_test(
        K,
        T,
        target_δ_rad,
        target_ψ_rad_deviation,
        time_second_interval,
        end_time_second,
        u0 = u0,
    )

ShipMMG.mmg_3dof_model! — Method

mmg_3dof_model!(dX, X, p, t)

MMG 3DOF model on DifferentialEquations.ODEProblem. Update dX.

Arguments

dX: [du, dv, dr, dδ, dnpm]

X: the initial state values. [u, v, r, δ, npm].

p: ρ and the basic & maneuvering parameters and δ & npm spline info.

ρ

L_pp

B

d

x_G

D_p

m

I_zG

A_R

η

m_x

m_y

J_z

f_α

ϵ

t_R

a_H

x_H

γRminus

γRplus

l_R

κ

t_P

w_P0

x_P

k_0

k_1

k_2

R0dash

Xvvdash

Xvrdash

Xrrdash

Xvvvvdash

Yvdash

Yrdash

Yvvvdash

Yvvrdash

Yvrrdash

Yrrrdash

Nvdash

Nrdash

Nvvvdash

Nvvrdash

Nvrrdash

Nrrrdash

spl_δ

spl_npm

t: the time.

ShipMMG.mmg_3dof_simulate — Method

mmg_3dof_simulate(time_list, npm_list, δ_list, basic_params, maneuvering_params, [, u0, v0, r0, ρ, algorithm, reltol, abstol]) -> time, u, v, r, δ, npm

Returns the MMG 3DOF simulation results including the lists of time, u, v, r, δ, npm. This function has the same logic of ShipMMG.simulate().

Arguments

time_list: the list of simulatino time.

npm_list: the list of propeller rpm.

δ_list: the list of rudder angle [rad].

basic_params::Mmg3DofBasicParams: the basic parameters of target ship.

maneuvering_params::Mmg3DofManeuveringParams: the maneuvering parameters of target ship.

u0::Float64=0.0: the initial x (surge) velocity.

v0::Float64=0.0: the initial y (sway) velocity.

r0::Float64=0.0: the initial rate of turn [rad/s].

ρ::Float64=1.025: the seawater density [g/cm^3].

algorithm=Tsit5(): the parameter of DifferentialEquations.ODEProblem.solve()

reltol=1e-8: the parameter of DifferentialEquations.ODEProblem.solve()

abstol=1e-8: the parameter of DifferentialEquations.ODEProblem.solve()

Examples

KVLCC2_L7 turning test.

julia> ρ = 1.025;
julia> L_pp = 7.00;
julia> B = 1.27;
julia> d = 0.46;
julia> nabla = 3.27;
julia> x_G = 0.25;
julia> # C_b = 0.810;
julia> D_p = 0.216;
julia> H_R = 0.345;
julia> A_R = 0.0539;
julia> t_P = 0.220;
julia> w_P0 = 0.40;
julia> m_x_dash = 0.022;
julia> m_y_dash = 0.223;
julia> J_z_dash = 0.011;
julia> t_R = 0.387;
julia> a_H = 0.312;
julia> x_H_dash = -0.464;
julia> γ_R_minus = 0.395;
julia> γ_R_plus = 0.640;
julia> l_r_dash = -0.710;
julia> x_P_dash = -0.690;
julia> ϵ = 1.09;
julia> κ = 0.50;
julia> f_α = 2.747;
julia> basic_params = Mmg3DofBasicParams();
julia> basic_params.L_pp = L_pp;  # 船長Lpp[m]
julia> basic_params.B = B;  # 船幅[m]
julia> basic_params.d = d;  # 喫水[m]
julia> basic_params.x_G = x_G;  # 重心位置[]
julia> basic_params.D_p = D_p;  # プロペラ直径[m]
julia> basic_params.m = ρ * nabla;  # 質量(無次元化)[kg]
julia> basic_params.I_zG = ρ * nabla * ((0.25 * L_pp)^2);  # 慣性モーメント[-]
julia> basic_params.A_R = A_R;  # 船の断面に対する舵面積比[-]
julia> basic_params.η = D_p / H_R;  # プロペラ直径に対する舵高さ(Dp/H)
julia> basic_params.m_x = (0.5 * ρ * (L_pp^2) * d) * m_x_dash;  # 付加質量x(無次元)
julia> basic_params.m_y = (0.5 * ρ * (L_pp^2) * d) * m_y_dash;  # 付加質量y(無次元)
julia> basic_params.J_z = (0.5 * ρ * (L_pp^4) * d) * J_z_dash;  # 付加質量Izz(無次元)
julia> basic_params.f_α = f_α; # 直圧力勾配係数
julia> basic_params.ϵ = ϵ;  # プロペラ・舵位置伴流係数比
julia> basic_params.t_R = t_R;  # 操縦抵抗減少率
julia> basic_params.a_H = a_H;  # 舵力増加係数
julia> basic_params.x_H = x_H_dash * L_pp;  # 舵力増分作用位置
julia> basic_params.γ_R_minus = γ_R_minus;  # 整流係数
julia> basic_params.γ_R_plus = γ_R_plus;  # 整流係数
julia> basic_params.l_R = l_r_dash;  # 船長に対する舵位置
julia> basic_params.κ = κ;  # 修正係数
julia> basic_params.t_P = t_P;  # 推力減少率
julia> basic_params.w_P0 = w_P0;  # 有効伴流率
julia> basic_params.x_P = x_P_dash;  # 船長に対するプロペラ位置
julia> maneuvering_params = Mmg3DofManeuveringParams();
julia> maneuvering_params.k_0 = 0.2931;
julia> maneuvering_params.k_1 = -0.2753;
julia> maneuvering_params.k_2 = -0.1385;
julia> maneuvering_params.R_0_dash = 0.022;
julia> maneuvering_params.X_vv_dash = -0.040;
julia> maneuvering_params.X_vr_dash = 0.002;
julia> maneuvering_params.X_rr_dash = 0.011;
julia> maneuvering_params.X_vvvv_dash = 0.771;
julia> maneuvering_params.Y_v_dash = -0.315;
julia> maneuvering_params.Y_r_dash = 0.083;
julia> maneuvering_params.Y_vvv_dash = -1.607;
julia> maneuvering_params.Y_vvr_dash = 0.379;
julia> maneuvering_params.Y_vrr_dash = -0.391;
julia> maneuvering_params.Y_rrr_dash = 0.008;
julia> maneuvering_params.N_v_dash = -0.137;
julia> maneuvering_params.N_r_dash = -0.049;
julia> maneuvering_params.N_vvv_dash = -0.030;
julia> maneuvering_params.N_vvr_dash = -0.294;
julia> maneuvering_params.N_vrr_dash = 0.055;
julia> maneuvering_params.N_rrr_dash = -0.013;
julia> duration = 200; # [s]
julia> max_δ_rad = 35 * pi / 180.0;  # [rad]
julia> n_const = 17.95;  # [rpm]
julia> sampling = duration * 10;
julia> time_list = range(0.00, stop = duration, length = sampling);
julia> δ_rad_list = max_δ_rad .* ones(Float64, sampling);
julia> npm_list = n_const .* ones(Float64, sampling);
julia> mmg_results = mmg_3dof_simulate(
    time_list,
    npm_list,
    δ_rad_list,
    basic_params,
    maneuvering_params,
    u0 = 2.29 * 0.512,
    v0 = 0.0,
    r0 = 0.0,
);

ShipMMG.mmg_3dof_zigzag_test — Method

mmg_3dof_zigzag_test(basic_params, maneuvering_params, npm_list, target_δ_rad, target_ψ_rad_deviation, time_second_interval, end_time_second, [, u0, v0, r0, ψ0, δ0, δ_rad_rate, start_time_second, algorithm, reltol, abstol]) -> final_δ_list, final_u_list, final_v_list, final_r_list, final_ψ_list

Returns the MMG 3DOF zigzag simulation results.

Arguments

basic_params::Mmg3DofBasicParams: the basic parameters of target ship.

maneuvering_params::Mmg3DofManeuveringParams: the maneuvering parameters of target ship.

npm_list: the list of propeller rpm.

target_δ_rad::Float64: target rudder angle of zigzag test.

target_ψ_rad_deviation::Float64: target azimuth deviation of zigzag test.

time_second_interval::Float64: time interval of output result.

end_time_second::Float64: the end time of simulation.

u0::Float64=0.0: the initial x (surge) velocity.

v0::Float64=0.0: the initial y (sway) velocity.

r0::Float64=0.0: the initial rate of turn [rad/s].

δ0::Float64=0.0: the initial rudder angle.

δ_rad_rate::Float64=10.0*π/180: the change rate of rudder angle [rad/s].

ρ::Float64=1.025: the seawater density [g/cm^3].

start_time_second::Float64=0.0: the start time of simulation.

algorithm=Tsit5(): the parameter of DifferentialEquations.ODEProblem.solve()

reltol=1e-8: the parameter of DifferentialEquations.ODEProblem.solve()

abstol=1e-8: the parameter of DifferentialEquations.ODEProblem.solve()

Examples

KVLCC2_L7 zigzag test.

julia> ρ = 1.025;
julia> L_pp = 7.00;
julia> B = 1.27;
julia> d = 0.46;
julia> nabla = 3.27;
julia> x_G = 0.25;
julia> # C_b = 0.810;
julia> D_p = 0.216;
julia> H_R = 0.345;
julia> A_R = 0.0539;
julia> t_P = 0.220;
julia> w_P0 = 0.40;
julia> m_x_dash = 0.022;
julia> m_y_dash = 0.223;
julia> J_z_dash = 0.011;
julia> t_R = 0.387;
julia> a_H = 0.312;
julia> x_H_dash = -0.464;
julia> γ_R_minus = 0.395;
julia> γ_R_plus = 0.640;
julia> l_r_dash = -0.710;
julia> x_P_dash = -0.690;
julia> ϵ = 1.09;
julia> κ = 0.50;
julia> f_α = 2.747;
julia> basic_params = Mmg3DofBasicParams();
julia> basic_params.L_pp = L_pp;  # 船長Lpp[m]
julia> basic_params.B = B;  # 船幅[m]
julia> basic_params.d = d;  # 喫水[m]
julia> basic_params.x_G = x_G;  # 重心位置[]
julia> basic_params.D_p = D_p;  # プロペラ直径[m]
julia> basic_params.m = ρ * nabla;  # 質量(無次元化)[kg]
julia> basic_params.I_zG = ρ * nabla * ((0.25 * L_pp)^2);  # 慣性モーメント[-]
julia> basic_params.A_R = A_R;  # 船の断面に対する舵面積比[-]
julia> basic_params.η = D_p / H_R;  # プロペラ直径に対する舵高さ(Dp/H)
julia> basic_params.m_x = (0.5 * ρ * (L_pp^2) * d) * m_x_dash;  # 付加質量x(無次元)
julia> basic_params.m_y = (0.5 * ρ * (L_pp^2) * d) * m_y_dash;  # 付加質量y(無次元)
julia> basic_params.J_z = (0.5 * ρ * (L_pp^4) * d) * J_z_dash;  # 付加質量Izz(無次元)
julia> basic_params.f_α = f_α; # 直圧力勾配係数
julia> basic_params.ϵ = ϵ;  # プロペラ・舵位置伴流係数比
julia> basic_params.t_R = t_R;  # 操縦抵抗減少率
julia> basic_params.a_H = a_H;  # 舵力増加係数
julia> basic_params.x_H = x_H_dash * L_pp;  # 舵力増分作用位置
julia> basic_params.γ_R_minus = γ_R_minus;  # 整流係数
julia> basic_params.γ_R_plus = γ_R_plus;  # 整流係数
julia> basic_params.l_R = l_r_dash;  # 船長に対する舵位置
julia> basic_params.κ = κ;  # 修正係数
julia> basic_params.t_P = t_P;  # 推力減少率
julia> basic_params.w_P0 = w_P0;  # 有効伴流率
julia> basic_params.x_P = x_P_dash;  # 船長に対するプロペラ位置
julia> maneuvering_params = Mmg3DofManeuveringParams();
julia> maneuvering_params.k_0 = 0.2931;
julia> maneuvering_params.k_1 = -0.2753;
julia> maneuvering_params.k_2 = -0.1385;
julia> maneuvering_params.R_0_dash = 0.022;
julia> maneuvering_params.X_vv_dash = -0.040;
julia> maneuvering_params.X_vr_dash = 0.002;
julia> maneuvering_params.X_rr_dash = 0.011;
julia> maneuvering_params.X_vvvv_dash = 0.771;
julia> maneuvering_params.Y_v_dash = -0.315;
julia> maneuvering_params.Y_r_dash = 0.083;
julia> maneuvering_params.Y_vvv_dash = -1.607;
julia> maneuvering_params.Y_vvr_dash = 0.379;
julia> maneuvering_params.Y_vrr_dash = -0.391;
julia> maneuvering_params.Y_rrr_dash = 0.008;
julia> maneuvering_params.N_v_dash = -0.137;
julia> maneuvering_params.N_r_dash = -0.049;
julia> maneuvering_params.N_vvv_dash = -0.030;
julia> maneuvering_params.N_vvr_dash = -0.294;
julia> maneuvering_params.N_vrr_dash = 0.055;
julia> maneuvering_params.N_rrr_dash = -0.013;
julia> target_δ_rad = 20.0 * π / 180.0
julia> target_ψ_rad_deviation = 20.0 * π / 180.0
julia> start_time_second = 0.00
julia> time_second_interval = 0.01
julia> end_time_second = 80.00
julia> time_list = start_time_second:time_second_interval:end_time_second
julia> n_const = 17.95  # [rpm]
julia> npm_list = n_const * ones(Float64, length(time_list))
julia> δ_list, u_list, v_list, r_list, ψ_list = mmg_3dof_zigzag_test(
    basic_params,
    maneuvering_params,
    npm_list,
    target_δ_rad,
    target_ψ_rad_deviation,
    time_second_interval,
    end_time_second,
);

ShipMMG.simulate — Method

mmg_3dof_simulate(time_list, npm_list, δ_list, L_pp, B, d, x_G, D_p, m, I_zG, A_R, η, m_x, m_y, J_z, f_α, ϵ, t_R, a_H, x_H, γ_R_minus, γ_R_plus, l_R, κ, t_P, w_P0, x_P, k_0, k_1, k_2, R_0_dash, X_vv_dash, X_vr_dash, X_rr_dash, X_vvvv_dash, Y_v_dash, Y_r_dash, Y_vvv_dash, Y_vvr_dash, Y_vrr_dash, Y_rrr_dash, N_v_dash, N_r_dash, N_vvv_dash, N_vvr_dash, N_vrr_dash, N_rrr_dash, [, u0, v0, r0, ρ, algorithm, reltol, abstol]) -> time, u, v, r, δ, npm

Returns the MMG 3DOF simulation results including the lists of time, u, v, r, δ, npm. This function has the same logic of ShipMMG.mmg_3dof_simulate().

Arguments

time_list: the list of simulatino time.

npm_list: the list of propeller rpm.

δ_list: the list of rudder angle [rad].

L_pp::Float64: L_pp

B::Float64:

d::Float64:

x_G::Float64:

D_p::Float64:

m::Float64:

I_zG::Float64:

A_R::Float64:

η::Float64:

m_x::Float64:

m_y::Float64:

J_z::Float64:

f_α::Float64:

ϵ::Float64:

t_R::Float64:

a_H::Float64:

x_H::Float64:

γ_R_minus::Float64:

γ_R_plus::Float64:

l_R::Float64:

κ::Float64:

t_P::Float64:

w_P0::Float64:

x_P::Float64:

k_0::Float64

k_1::Float64

k_2::Float64

R_0_dash::Float64

X_vv_dash::Float64

X_vr_dash::Float64

X_rr_dash::Float64

X_vvvv_dash::Float64

Y_v_dash::Float64

Y_r_dash::Float64

Y_vvv_dash::Float64

Y_vvr_dash::Float64

Y_vrr_dash::Float64

Y_rrr_dash::Float64

N_v_dash::Float64

N_r_dash::Float64

N_vvv_dash::Float64

N_vvr_dash::Float64

N_vrr_dash::Float64

N_rrr_dash::Float64

u0::Float64=0.0: the initial x (surge) velocity.

v0::Float64=0.0: the initial y (sway) velocity.

r0::Float64=0.0: the initial rate of turn [rad/s].

ρ::Float64=1.025: the seawater density [g/cm^3].

algorithm=Tsit5(): the parameter of DifferentialEquations.ODEProblem.solve()

reltol=1e-8: the parameter of DifferentialEquations.ODEProblem.solve()

abstol=1e-8: the parameter of DifferentialEquations.ODEProblem.solve()